feat: 메인 배너 컴포넌트 화 및 메인 페이지로 이동
